+++ code/crates/test/app/src/node.rs
@@ -254,7 +254,7 @@ fn make_config(index: usize, total: usize, settings: MakeConfigSettings) -> Conf
             enabled: true,
             // Current test app does not support proposal-only value payload properly as Init does not include valid_round
             value_payload: ValuePayload::ProposalAndParts,
-            queue_capacity: 100, // Deprecated, derived from `sync.parallel_requests`
+            queue_capacity: 100,
             p2p: P2pConfig {
                 protocol: PubSubProtocol::default(),
                 listen_addr: settings.transport.multiaddr("127.0.0.1", consensus_port),